+++ modules/flopy/modflow/mfdrt.py
@@ -172,16 +172,16 @@ class ModflowDrt(Package):
     @staticmethod
     def get_default_dtype(structured=True):
         if structured:
-            dtype = np.dtype([("k", np.int), ("i", np.int),
-                              ("j", np.int), ("elev", np.float32),
-                              ("cond", np.float32), ("layr", np.int),
-                              ("rowr", np.int), ("colr", np.int),
-                              ("rfprop", np.float32)])
+            dtype = np.dtype([("k", np.int_), ("i", np.int_),
+                              ("j", np.int_), ("elev", np.float64),
+                              ("cond", np.float64), ("layr", np.int_),
+                              ("rowr", np.int_), ("colr", np.int_),
+                              ("rfprop", np.float64)])
         else:
-            dtype = np.dtype([("inode", np.int), ("elev", np.float32),
-                              ("cond", np.float32), ("layr", np.int),
-                              ("rowr", np.int), ("colr", np.int),
-                              ("rfprop", np.float32)])
+            dtype = np.dtype([("inode", np.int_), ("elev", np.float64),
+                              ("cond", np.float64), ("layr", np.int_),
+                              ("rowr", np.int_), ("colr", np.int_),
+                              ("rfprop", np.float64)])
         return dtype
 
     def ncells(self):
@@ -229,7 +229,7 @@ class ModflowDrt(Package):
         # get an empty recarray that corresponds to dtype
         dtype = ModflowDrt.get_default_dtype(structured=structured)
         if aux_names is not None:
-            dtype = Package.add_to_dtype(dtype, aux_names, np.float32)
+            dtype = Package.add_to_dtype(dtype, aux_names, np.float64)
         return create_empty_recarray(ncells, dtype, default_value=-1.0E+10)
 
     @staticmethod
